Extract code into method HelloApplication to be more readable.
This simplifies 6 lines of code to help make run() more readable. Simplified to a method described in 3 words and a number representing how many milliseconds to wait to get data after. This prevents needing to look through the lines to figure out how run() works. I kept it as run() because it is simple and direct, if more information is needed then getDataAfter(millis) explains the contents faster. Specifying it is in millis seems like too much information as the code editior displays the variable name there too, and millis has been the default time delay unit so it should be fair to not explicitly mention compared to hours or minutes.

getDataAfter(int millis)
getDataAfter(1000)