Simplify and clean of repeating of code in HCTimeSlotTest.
Since we were repeating the same 2 lines over and over with only changing two variables it seemed a lot easier to combine them into one action with two variables. This helps get rid of the repeating code smell and makes it easier to adjust the two parts that are changing each time.