Renamed the clear drop target signal in DragAndDropController again, (I won't count this change as toward the count because it is only updating a previous refactor).
I found it unnecessary to specify that it should be valid, removing that word makes it clear what the action that method does. The validity does not matter. This line of code would still work even if we had it highlighting red as an invalid drop target for example. By fixing this we are clearing up potential assumptions that might be made about it that are wrong. All it does is clear the signal back to default, and now the name reflects that.