Extracted the repeating try catch that occurred across most of the Classes in elements folder.
This extraction summarizes around 4 lines of code into the words tryLoad, and places it in a new TryLoadHolder Class.
Since unnecessary repeats of code is a code smell, I made a new Class to hold the method. Now when those lines of code need changed we wouldn't need to hunt through the elements folder to find where else it is used. As all those points now simply reference this Class.
There was one variation of the lines of code. To deal with that I added a second method with a longer name to handle that instance to have it in the same place.